cloud_functions/hibob/holiday_balances/pyproject.toml
[tool.poetry]
name = "holiday balances"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
data_pipeline_tools = { git = "https://github.com/tpximpact/GCP-Data-Pipelines.git", branch = "main" }
requests = "^2.28.2"
google-cloud-core = "^2.3.2"
google-cloud-bigquery = "^3.5.0"
google-cloud-secret-manager = "^2.15.1"
pandas-gbq = "^0.19.1"
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
